Move most methods in ApiUtils to smaller, more specific util classes
