+++ src/main/java/blusunrize/immersiveengineering/api/crafting/IngredientWithSize.java
@@ -9,23 +9,18 @@
 
 package blusunrize.immersiveengineering.api.crafting;
 
-import blusunrize.immersiveengineering.api.ApiUtils;
+import blusunrize.immersiveengineering.api.utils.ItemUtils;
 import blusunrize.immersiveengineering.common.crafting.IngredientWithSizeSerializer;
 import com.google.gson.JsonElement;
-import com.google.gson.JsonObject;
-import it.unimi.dsi.fastutil.ints.IntList;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.crafting.Ingredient;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.tags.Tag;
-import net.minecraftforge.common.crafting.IIngredientSerializer;
 
 import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
-import java.util.List;
 import java.util.function.Predicate;
-import java.util.stream.Stream;
 
 public class IngredientWithSize implements Predicate<ItemStack>
 {
@@ -77,7 +72,7 @@ public class IngredientWithSize implements Predicate<ItemStack>
 		ItemStack[] baseStacks = basePredicate.getMatchingStacks();
 		ItemStack[] ret = new ItemStack[baseStacks.length];
 		for(int i = 0; i < baseStacks.length; ++i)
-			ret[i] = ApiUtils.copyStackWithAmount(baseStacks[i], this.count);
+			ret[i] = ItemUtils.copyStackWithAmount(baseStacks[i], this.count);
 		return ret;
 	}
