+++ src/main/java/blusunrize/immersiveengineering/common/blocks/generic/PoweredMultiblockTileEntity.java
@@ -10,12 +10,12 @@ package blusunrize.immersiveengineering.common.blocks.generic;
 
 import blusunrize.immersiveengineering.api.ApiUtils;
 import blusunrize.immersiveengineering.api.IEEnums.IOSideConfig;
-import blusunrize.immersiveengineering.api.crafting.IMultiblockRecipe;
 import blusunrize.immersiveengineering.api.crafting.IngredientWithSize;
 import blusunrize.immersiveengineering.api.crafting.MultiblockRecipe;
 import blusunrize.immersiveengineering.api.energy.immersiveflux.FluxStorage;
 import blusunrize.immersiveengineering.api.energy.immersiveflux.FluxStorageAdvanced;
 import blusunrize.immersiveengineering.api.multiblocks.TemplateMultiblock;
+import blusunrize.immersiveengineering.api.utils.IngredientUtils;
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.IComparatorOverride;
 import blusunrize.immersiveengineering.common.blocks.IEBlockInterfaces.IProcessTile;
 import blusunrize.immersiveengineering.common.blocks.multiblocks.IETemplateMultiblock;
@@ -24,7 +24,6 @@ import blusunrize.immersiveengineering.common.util.EnergyHelper.IIEInternalFluxH
 import blusunrize.immersiveengineering.common.util.Utils;
 import blusunrize.immersiveengineering.common.util.inventory.IIEInventory;
 import net.minecraft.item.ItemStack;
-import net.minecraft.item.crafting.IRecipe;
 import net.minecraft.nbt.CompoundNBT;
 import net.minecraft.nbt.ListNBT;
 import net.minecraft.tileentity.TileEntityType;
@@ -617,7 +616,7 @@ public abstract class PoweredMultiblockTileEntity<T extends PoweredMultiblockTil
 				for(int i = 0; i < inputSlots.length; i++)
 					if(inputSlots[i] >= 0&&inputSlots[i] < inv.size())
 						query.set(i, multiblock.getInventory().get(inputSlots[i]));
-				if(!ApiUtils.stacksMatchIngredientWithSizeList(recipe.getItemInputs(), query))
+				if(!IngredientUtils.stacksMatchIngredientWithSizeList(recipe.getItemInputs(), query))
 				{
 					this.clearProcess = true;
 					return;
