+++ src/main/java/blusunrize/immersiveengineering/common/items/VoltmeterItem.java
@@ -19,25 +19,21 @@ import blusunrize.immersiveengineering.api.wires.IImmersiveConnectable;
 import blusunrize.immersiveengineering.api.wires.LocalWireNetwork;
 import blusunrize.immersiveengineering.api.wires.localhandlers.EnergyTransferHandler;
 import blusunrize.immersiveengineering.api.wires.localhandlers.EnergyTransferHandler.Path;
+import blusunrize.immersiveengineering.api.wires.utils.WirecoilUtils;
+import blusunrize.immersiveengineering.api.wires.utils.WirecoilUtils.WireLink;
 import blusunrize.immersiveengineering.common.util.ChatUtils;
-import blusunrize.immersiveengineering.common.util.ItemNBTHelper;
 import blusunrize.immersiveengineering.common.util.Utils;
-import jdk.nashorn.internal.runtime.regexp.joni.constants.TargetInfo;
 import net.minecraft.client.util.ITooltipFlag;
 import net.minecraft.entity.player.PlayerEntity;
 import net.minecraft.item.ItemStack;
 import net.minecraft.item.ItemUseContext;
-import net.minecraft.nbt.CompoundNBT;
-import net.minecraft.nbt.NBTUtil;
 import net.minecraft.tileentity.TileEntity;
 import net.minecraft.util.ActionResultType;
 import net.minecraft.util.Direction;
 import net.minecraft.util.math.BlockPos;
-import net.minecraft.util.math.Vec3i;
 import net.minecraft.util.text.ITextComponent;
 import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraft.world.World;
-import org.apache.http.pool.ConnPool;
 
 import javax.annotation.Nullable;
 import java.util.List;
@@ -53,12 +49,16 @@ public class VoltmeterItem extends IEBaseItem implements ITool
 	public void addInformation(ItemStack stack, @Nullable World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn)
 	{
 		super.addInformation(stack, worldIn, tooltip, flagIn);
-		if(ItemNBTHelper.hasKey(stack, "linkingPos"))
+		if(WirecoilUtils.hasWireLink(stack))
 		{
-			CompoundNBT link = stack.getOrCreateTag().getCompound("linkingPos");
-			String dimension = stack.getOrCreateTag().getString("linkingDim");
-			ConnectionPoint cp = new ConnectionPoint(link);
-			tooltip.add(new TranslationTextComponent(Lib.DESC_INFO+"attachedToDim", cp.getX(), cp.getY(), cp.getZ(), dimension));
+			WirecoilUtils.WireLink link = WirecoilUtils.WireLink.readFromItem(stack);
+			tooltip.add(new TranslationTextComponent(
+					Lib.DESC_INFO+"attachedToDim",
+					link.cp.getX(),
+					link.cp.getY(),
+					link.cp.getZ(),
+					link.dimension.toString()
+			));
 		}
 	}
 
@@ -79,8 +79,8 @@ public class VoltmeterItem extends IEBaseItem implements ITool
 		TileEntity tileEntity = world.getTileEntity(pos);
 		if((player==null||!player.isSneaking())&&(tileEntity instanceof IFluxReceiver||tileEntity instanceof IFluxProvider))
 		{
-			int max = 0;
-			int stored = 0;
+			int max;
+			int stored;
 			if(tileEntity instanceof IFluxReceiver)
 			{
 				max = ((IFluxReceiver)tileEntity).getMaxEnergyStored(side);
@@ -101,32 +101,30 @@ public class VoltmeterItem extends IEBaseItem implements ITool
 				return ActionResultType.SUCCESS;
 			TargetingInfo targetingInfo = new TargetingInfo(context);
 			BlockPos masterPos = ((IImmersiveConnectable)tileEntity).getConnectionMaster(null, targetingInfo);
-			Vec3i delta = pos.subtract(masterPos);
+			BlockPos delta = pos.subtract(masterPos);
 			ConnectionPoint cp = ((IImmersiveConnectable)tileEntity).getTargetedPoint(targetingInfo, delta);
 			if(cp==null)
 				return ActionResultType.FAIL;
-			if(!ItemNBTHelper.hasKey(stack, "linkingPos"))
+			if(WirecoilUtils.hasWireLink(stack))
 			{
-				CompoundNBT nbt = stack.getOrCreateTag();
-				nbt.putString("linkingDim", world.getDimension().getType().getRegistryName().toString());
-				nbt.put("linkingPos", cp.createTag());
+				WireLink link = WireLink.create(cp, world, delta, targetingInfo);
+				link.writeToItem(stack);
 			}
 			else
 			{
-				String dim = ItemNBTHelper.getString(stack, "linkingDim");
-				if(dim.equals(world.getDimension().getType().getRegistryName().toString()))
+				WireLink link = WireLink.readFromItem(stack);
+				if(link.dimension.equals(world.getDimension().getType().getRegistryName()))
 				{
-					ConnectionPoint linkCP = new ConnectionPoint(stack.getOrCreateTag().getCompound("linkingPos"));
 					GlobalWireNetwork global = GlobalWireNetwork.getNetwork(world);
 					LocalWireNetwork netHere = global.getNullableLocalNet(cp);
-					LocalWireNetwork netLink = global.getNullableLocalNet(linkCP);
+					LocalWireNetwork netLink = global.getNullableLocalNet(link.cp);
 					if(netHere==netLink&&netHere!=null)
 					{
 						EnergyTransferHandler energyHandler = netHere.getHandler(EnergyTransferHandler.ID,
 								EnergyTransferHandler.class);
 						if(energyHandler!=null)
 						{
-							Path energyPath = energyHandler.getPath(linkCP, cp);
+							Path energyPath = energyHandler.getPath(link.cp, cp);
 							if(energyPath!=null)
 								player.sendMessage(new TranslationTextComponent(
 										Lib.CHAT_INFO+"averageLoss",
@@ -135,8 +133,7 @@ public class VoltmeterItem extends IEBaseItem implements ITool
 						}
 					}
 				}
-				ItemNBTHelper.remove(stack, "linkingPos");
-				ItemNBTHelper.remove(stack, "linkingDim");
+				WirecoilUtils.clearWireLink(stack);
 			}
 			return ActionResultType.SUCCESS;
 		}
